[Lime][Orange][Kumquat] fix bug: crash when F9 from tangerine
Tag of root node of each deserialized hierarchy was set to asset bundle
path of this scene. But not in Tangerine to avoid confusing users
and also not to serialize it. CodeCooker used to depend on the fact
that there's always path to scene in a root's tag. Which lead to crash
in Tangerine.

So this commit introduces new component for storing this data always
leaving tag empty.