[CIT-1392][Tangerine] Remove FocusScope from DocumentArea
Focus scope on docArea is useless due to its content (there're only one tab traversal field - zoom), but moreover it produces undefined behavior and bugs in exposition mode due to Key.Tab repeated state. As a result Key.Tab behavior and its repeated state depends on order of processing node's updates. Another solution is to consume Key.Tab in exposition mode Task, but it's not quite good because it's not obvious and looks like a "hack".